Added CQRS approach in product classes
