+++ src/main/java/es/cloudapps/hexarch/hexagon/application/impl/ProductQueryService.java
@@ -1,6 +1,6 @@
 package es.cloudapps.hexarch.hexagon.application.impl;
 
-import es.cloudapps.hexarch.hexagon.application.ProductServicePort;
+import es.cloudapps.hexarch.hexagon.application.ProductQueryServicePort;
 import es.cloudapps.hexarch.hexagon.domain.exception.NotFoundException;
 import es.cloudapps.hexarch.hexagon.domain.model.Product;
 import es.cloudapps.hexarch.hexagon.domain.services.ProductRepository;
@@ -8,27 +8,12 @@ import es.cloudapps.hexarch.hexagon.domain.services.ProductRepository;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class ProductService implements ProductServicePort {
+public class ProductQueryService implements ProductQueryServicePort {
 
     private final ProductRepository productRepository;
 
-    public ProductService(ProductRepository productRepository) {
+    public ProductQueryService(ProductRepository productRepository) {
         this.productRepository = productRepository;
-    }
-
-    @Override
-    public RegisterNewResp registerNew(RegisterNewReq params) {
-        Product product = new Product(new Product.Name(params.name));
-        Product saved = productRepository.save(product);
-        return new RegisterNewResp(saved.id(), saved.name());
-    }
-
-    @Override
-    public RemoveResp remove(RemoveReq params) {
-        Product existing = productRepository.get(params.id).orElseThrow(NotFoundException::new);
-        productRepository.remove(existing);
-
-        return new RemoveResp();
     }
 
     @Override
