+++ src/main/java/es/cloudapps/hexarch/infrastructure/di/Injector.java
@@ -1,10 +1,12 @@
 package es.cloudapps.hexarch.infrastructure.di;
 
 import es.cloudapps.hexarch.hexagon.application.CartExpenditureServicePort;
-import es.cloudapps.hexarch.hexagon.application.ProductServicePort;
+import es.cloudapps.hexarch.hexagon.application.ProductCommandServicePort;
+import es.cloudapps.hexarch.hexagon.application.ProductQueryServicePort;
 import es.cloudapps.hexarch.hexagon.application.ShoppingCartServicePort;
 import es.cloudapps.hexarch.hexagon.application.impl.CartExpenditureService;
-import es.cloudapps.hexarch.hexagon.application.impl.ProductService;
+import es.cloudapps.hexarch.hexagon.application.impl.ProductCommandService;
+import es.cloudapps.hexarch.hexagon.application.impl.ProductQueryService;
 import es.cloudapps.hexarch.hexagon.application.impl.ShoppingCartService;
 import es.cloudapps.hexarch.hexagon.domain.services.CheckoutService;
 import es.cloudapps.hexarch.hexagon.domain.services.ProductRepository;
@@ -16,8 +18,13 @@ import org.springframework.context.annotation.Configuration;
 public class Injector {
 
     @Bean
-    public ProductServicePort productServicePort(ProductRepository repository) {
-        return new ProductService(repository);
+    public ProductQueryServicePort productQueryServicePort(ProductRepository repository) {
+        return new ProductQueryService(repository);
+    }
+
+    @Bean
+    public ProductCommandServicePort productCommandServicePort(ProductRepository repository){
+        return new ProductCommandService(repository);
     }
 
     @Bean
