+++ src/test/java/es/cloudapps/hexarch/hexagon/application/impl/ProductServiceTest.java
@@ -1,6 +1,7 @@
 package es.cloudapps.hexarch.hexagon.application.impl;
 
-import es.cloudapps.hexarch.hexagon.application.ProductServicePort.*;
+import es.cloudapps.hexarch.hexagon.application.ProductCommandServicePort.*;
+import es.cloudapps.hexarch.hexagon.application.ProductQueryServicePort.*;
 import es.cloudapps.hexarch.hexagon.domain.exception.NotFoundException;
 import es.cloudapps.hexarch.hexagon.domain.model.Product;
 import es.cloudapps.hexarch.hexagon.domain.services.ProductRepository;
@@ -20,12 +21,14 @@ public class ProductServiceTest {
 
     private ProductRepository repository;
 
-    private ProductService productService;
+    private ProductQueryService productService;
+    private ProductCommandService productCommandService;
 
     @Before
     public void setUp() {
         this.repository = mock(ProductRepository.class);
-        this.productService = new ProductService(repository);
+        this.productService = new ProductQueryService(repository);
+        this.productCommandService = new ProductCommandService(repository);
     }
 
     @Test
@@ -39,9 +42,9 @@ public class ProductServiceTest {
 
         when(repository.save(any())).thenReturn(product);
 
-        RegisterNewResp resp = productService.registerNew(req);
+        RegisterNewResp resp = productCommandService.registerNew(req);
 
-        verify(repository,times(1)).save(any());
+        verify(repository, times(1)).save(any());
 
         assertNotNull(resp);
         assertThat(resp.name, is(equalTo("Test product")));
@@ -57,9 +60,9 @@ public class ProductServiceTest {
 
         when(repository.get(1)).thenReturn(Optional.of(product));
 
-        RemoveResp resp = productService.remove(req);
+        RemoveResp resp = productCommandService.remove(req);
 
-        verify(repository,times(1)).remove(product);
+        verify(repository, times(1)).remove(product);
 
         assertNotNull(resp);
     }
@@ -71,7 +74,7 @@ public class ProductServiceTest {
 
         when(repository.get(any())).thenReturn(Optional.empty());
 
-        productService.remove(req);
+        productCommandService.remove(req);
     }
 
     @Test(expected = NotFoundException.class)
@@ -92,7 +95,7 @@ public class ProductServiceTest {
         Product p2 = mock(Product.class);
         Product p3 = mock(Product.class);
 
-        when(repository.all()).thenReturn(Arrays.asList(p1,p2,p3));
+        when(repository.all()).thenReturn(Arrays.asList(p1, p2, p3));
 
         FindAllResp resp = productService.findAll(req);
